refactor: added shop file to cogshop model
