+++ cognite/powerops/resync/models/cdf_resources.py
