+++ cognite/powerops/resync/models/market/rkom.py
@@ -6,8 +6,8 @@ from pydantic import Field
 
 from cognite.powerops.cdf_labels import AssetLabel
 from cognite.powerops.resync.models._base import AssetType, NonAssetType
+from cognite.powerops.resync.models.cdf_resources import CDFSequence
 from cognite.powerops.resync.models.market.base import Bid, Market, Process
-from cognite.powerops.resync.models.shared import CDFSequence
 
 
 class RKOMBid(Bid):
